# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Maddy Backend API
  version: 0.1.0
  description: Offline-first sync backend for Maddy iPhone and macOS clients.
servers:
  - url: http://localhost:4000/v1
paths:
  /health/live:
    get:
      summary: Liveness check
      responses:
        '200':
          description: OK
  /health/ready:
    get:
      summary: Readiness check
      responses:
        '200':
          description: Ready
  /version:
    get:
      summary: Version info
      responses:
        '200':
          description: Version payload
  /tasks:
    get:
      summary: List tasks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Task list
    post:
      summary: Create task
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Created
  /tasks/{id}:
    patch:
      summary: Update task
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Updated
    delete:
      summary: Soft-delete task
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Deleted
  /sync/push:
    post:
      summary: Push client mutations
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Push accepted
  /sync/pull:
    get:
      summary: Pull delta changes by cursor
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: cursor
          schema:
            type: string
      responses:
        '200':
          description: Delta changes
  /sync/ack:
    post:
      summary: Ack cursor for device
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Acked
  /attachments/presign-upload:
    post:
      summary: Get presigned upload URL
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Presigned URL
  /attachments/{id}/complete:
    post:
      summary: Mark upload complete
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Completed
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT